move libpng engine shits over to a libpng extension
