+++ ENIGMAsystem/SHELL/Universal_System/Extensions/libpng/Makefile
@@ -1,0 +1,5 @@
+SHARED_SRC_DIR := ../../shared
+override CXXFLAGS += -I$(SHARED_SRC_DIR)/libpng-util -I$(SHARED_SRC_DIR)
+override LDLIBS += -L$(SHARED_SRC_DIR)/libpng-util -lpng-util -lpng -lz
+
+SOURCES += Universal_System/Extensions/libpng/libpng-ext.cpp
