+++ ENIGMAsystem/SHELL/Universal_System/Extensions/libpng/libpng-ext.h
@@ -1,0 +1,27 @@
+/** Copyright (C) 2019 Robert B. Colton
+***
+*** This file is a part of the ENIGMA Development Environment.
+***
+*** ENIGMA is free software: you can redistribute it and/or modify it under the
+*** terms of the GNU General Public License as published by the Free Software
+*** Foundation, version 3 of the license or any later version.
+***
+*** This application and its source code is distributed AS-IS, WITHOUT ANY
+*** WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+*** FOR A PARTICULAR PURPOSE. See the GNU General Public License for more
+*** details.
+***
+*** You should have received a copy of the GNU General Public License along
+*** with this code. If not, see <http://www.gnu.org/licenses/>
+**/
+
+#ifndef ENIGMA_LIBPNG_H
+#define ENIGMA_LIBPNG_H
+
+namespace enigma {
+
+void extension_libpng_init();
+
+} // namespace enigma
+
+#endif // ENIGMA_LIBPNG_H
