+++ ENIGMAsystem/SHELL/Universal_System/image_formats.h
@@ -19,12 +19,21 @@
 #ifndef ENIGMA_IMAGEFORMATS_H
 #define ENIGMA_IMAGEFORMATS_H
 
+#include <map>
+#include <functional>
 #include <string>
+
+using ImageLoadFunction = std::function<unsigned char*(std::string, unsigned int*, unsigned int*, unsigned int*, unsigned int*, bool)>;
+using ImageSaveFunction = std::function<int(std::string, const unsigned char*, unsigned, unsigned, unsigned, unsigned, bool)>;
 
 /// NOTE: These image formats expect the data to be un-aligned and always reads and writes with BGRA full color
 
-namespace enigma 
+namespace enigma
 {
+
+extern std::map<std::string, ImageLoadFunction> image_load_handlers;
+extern std::map<std::string, ImageSaveFunction> image_save_handlers;
+
 /// Color formats
 enum {
   color_fmt_rgba,
@@ -48,10 +57,9 @@ int image_save(std::string filename, const unsigned char* data, std::string form
 int image_save(std::string filename, const unsigned char* data, unsigned width, unsigned height, unsigned fullwidth, unsigned fullheight, bool flipped);
 
 unsigned char* image_load_bmp(std::string filename, unsigned int* width, unsigned int* height, unsigned int* fullwidth, unsigned int* fullheight, bool flipped);
-unsigned char* image_load_png(std::string filename, unsigned int* width, unsigned int* height, unsigned int* fullwidth, unsigned int* fullheight, bool flipped);
 unsigned char* image_load_gif(std::string filename, unsigned int* width, unsigned int* height, unsigned int* fullwidth, unsigned int* fullheight, int* imgnumb, bool flipped);
 int image_save_bmp(std::string filename, const unsigned char* data, unsigned width, unsigned height, unsigned fullwidth, unsigned fullheight, bool flipped);
-int image_save_png(std::string filename, const unsigned char* data, unsigned width, unsigned height, unsigned fullwidth, unsigned fullheight, bool flipped);
+
 } //namespace enigma
 
 #endif //ENIGMA_IMAGEFORMATS_H
